Backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
joblib==1.5.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
